[FEATURE] Support for quoted keys in arrays
The syntax for Fluid arrays is based on the JSON syntax, but up to now
it didn't support quoted arrays, so::

 {x:someViewHelper(arrayArgument: '{"foo": "bar"}')}

wouldn't parse.

With this change the parser is a little less strict for arrays allowing
the keys to be quoted with single or double quotes.
It also allows to use special characters within quoted keys::

 <x:someViewHelper arrayArgument="{'@this[will]': 'work'}" />
 {x:someViewhelper(arrayArgument: '{"and.so": "will.this"}'}

See https://jira.neos.io/browse/FLOW-248 for more information.