[BUGFIX] Always wrap ViewHelper arguments in SyntaxTree Nodes
This fixes a regression that prevents usage of simple types for
boolean ViewHelper arguments.

Without this fix, Fluid code like the following lead to fatal
PHP errors::

 {f:if(condition: 1, then: 'should be rendered')}

Because Fluid tries to instantiate a ``BooleanNode`` without creating
the root node.

In tag syntax the argument is converted to a ``TextNode``, so the
following works without this patch::

 <f:if condition="1" then="should be rendered" />

Note: This also tweaks the functional tests to actually verify the
results of the *Conditions* test cases.